requirements.txt
Django==2.2.12
django-braces==1.11.0
Pillow==6.2.2